pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clippyshot"
version = "0.1.0"
description = "Sandboxed LibreOffice document-to-image renderer with perceptual hashing"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "magika>=0.5.1",
    "ImageHash>=4.3.1",
    "Pillow>=11.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "psycopg[binary]>=3.2.0",
    "redis>=5.0.0",
    "pypdf>=4.0.0",
    "pyzipper>=0.3.6",
    "python-magic>=0.4.27",
    "olefile>=0.46",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "fakeredis>=2.21.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[project.scripts]
clippyshot = "clippyshot.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
clippyshot = ["static/*.html", "static/*.css", "static/*.js", "static/assets/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-ra -q"
markers = [
    "integration: requires real soffice and sandbox backend (slow)",
    "docker: requires docker daemon",
]
asyncio_mode = "auto"
